requirements.txt
djangorestframework
markdown
django-filter
drf-extra-fields
pygments
djangorestframework-simplejwt
djoser
beautifulsoup4
nh3
drf-nested-routers
python-dotenv